Move integration test utils to central package
